# openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User API",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "Users",
      "description": "Operasi seputar user"
    }
  ],
  "paths": {
    "/api/users": {
      "post": {
        "tags": ["Users"],
        "summary": "Create new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created user"
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "summary": "Get current user info",
        "responses": {
          "200": {
            "description": "User info returned"
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Logout current user",
        "responses": {
          "200": {
            "description": "User logged out"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login user and send verification email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification code sent"
          }
        }
      }
    },
    "/api/users/verification-code/welcome/{verifCode}": {
      "get": {
        "tags": ["Users"],
        "summary": "Verify user account via token",
        "parameters": [
          {
            "in": "path",
            "name": "verifCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful verification"
          }
        }
      }
    },
    "/api/users/recovery": {
      "post": {
        "tags": ["Users"],
        "summary": "Send password recovery code",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/emailDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Recovery token sent"
          }
        }
      }
    },
    "/api/users/recovery/password": {
      "post": {
        "tags": ["Users"],
        "summary": "Update user password after recovery",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Get user by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateUserDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "email", "password"]
      },
      "LoginUserDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": ["username", "password"]
      },
      "emailDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": ["email"]
      },
      "PasswordDTO": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        },
        "required": ["password"]
      }
    }
  }
}